GPT validation and OPENAI Utility
updated imports for relative import using ".".
Also added find users and get all data to users.
Find users can be used to find user based on different criterions in the database.
Finally added the validation for output from chatgpt and then added validation for the output in the Utility file